Allow searching in HTML report via URL search param (#1302)
* Prevent SearchBar from triggering HTML form action

This commit stops the `SearchBar` from triggering an HTML form action
when the search button is clicked. In addition, it correctly changes the
`SearchQueryContext` such that other components (e.g. `FilteredResults`
and `Highlight`) can react to the change, rather than using a
state-setter passed down via props.

* Add tests

* Lint fix

* Update CHANGELOG

* Bind search to URL search param

In the `react` library, the `SearchQuery` can now be optionally bound to
a URL search parameter in the `window.location`. If this option is
enabled, then when the search query context is updated, it causes the
page to reload with an updated search string, rather than updating the
context's state internally.

In `html-formatter`, this new option is used, bound to the 'search'
URL search parameter.

This makes it possible to create hyperlinks to a web page of cucumber
results with a particular search applied. This could be used in DevOps
workflows, for example by attaching a link to an issue ticket that
points to a set of features tagged with the issue number.

* Update changelog

* Remove unnecessary hidden input component in form

* Correct link in changelog

* Another changelog fix

* Rework to only link one-way

* Lint fixes

* Remove unnecessary export

* fix import and enum

* readd allStatuses array

* lint

* remove duplicate statuses array

* remove redundant sort

`statuses` is ordered already

* destructure for clarity

* Use `history.replaceState` instead of showing link

* Fix bug with rendering gherkin samples

* Inverse filter parameter from 'hide'->'only'

* Revert to `hide=` for status filters

* remove --exit

* Update html-formatter/CHANGELOG.md

Co-authored-by: Aurélien Reeves <aurelien.reeves@smartbear.com>

Co-authored-by: Aslak Hellesøy <1000+aslakhellesoy@users.noreply.github.com>
Co-authored-by: davidgoss <david@davidgoss.co>
Co-authored-by: Aurélien Reeves <aurelien.reeves@smartbear.com>